requirements.txt
streamlit==1.35.0
torch==2.2.0
transformers==4.35.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
matplotlib==3.8.2
scikit-image==0.22.0